Day 11: seating layout
I got stuck for a while on part 1: while swapping layouts, I was
forgetting to put the previous value in the new grid for the default
case, so it would just alternate between fully seated and no one. Oops.

In the second part, I had the logic mostly right but had to do a bunch
of debugging to realize I had mixed my indices and was getting
`layout[i][j]` when I wanted to get `layout[j][i]`.